requirements.txt
mkdocs-material>=9.6.0
mkdocs-jupyter>=0.25.0
